refactor: Move routes to /api (#149)
