+++ src/api/routes/synapse/room_events.rs
@@ -10,6 +10,7 @@ use sqlx::{Postgres, Transaction};
 use uuid::Uuid;
 
 use crate::{
+    api::routes::v1::error::CommonError,
     components::{
         app::AppComponents,
         database::{DatabaseComponent, DatabaseComponentImplementation},
@@ -20,7 +21,6 @@ use crate::{
         friendships::{Friendship, FriendshipRepositoryImplementation, FriendshipsRepository},
     },
     middlewares::check_auth::{Token, UserId},
-    routes::v1::error::CommonError,
 };
 
 use super::errors::SynapseError;
@@ -279,7 +279,7 @@ async fn get_friendship_from_db(
 
         log::warn!("Error getting friendship in room event {}", err);
         return Err(SynapseError::CommonError(
-            crate::routes::v1::error::CommonError::Unknown,
+            crate::api::routes::v1::error::CommonError::Unknown,
         ));
     }
 
@@ -306,7 +306,7 @@ async fn get_last_history_from_db(
 
         log::warn!("Error getting friendship history in room event {}", err);
         return Err(SynapseError::CommonError(
-            crate::routes::v1::error::CommonError::Unknown,
+            crate::api::routes::v1::error::CommonError::Unknown,
         ));
     }
 
@@ -551,7 +551,7 @@ mod tests {
     use chrono::NaiveDate;
 
     use crate::{
-        entities::friendship_history::FriendshipHistory, routes::synapse::errors::SynapseError,
+        api::routes::synapse::errors::SynapseError, entities::friendship_history::FriendshipHistory,
     };
 
     use super::{process_friendship_status, FriendshipEvent, FriendshipStatus};
