+++ src/api/routes/v1/friendships/mutuals.rs
@@ -6,10 +6,10 @@ use actix_web::{
 
 use super::{errors::FriendshipsError, types::FriendshipsResponse};
 use crate::{
+    api::routes::v1::error::CommonError,
     components::{app::AppComponents, synapse::clean_synapse_user_id},
     entities::friendships::FriendshipRepositoryImplementation,
     middlewares::check_auth::UserId,
-    routes::v1::error::CommonError,
 };
 
 #[get("/v1/friendships/{userId}/mutuals")]
